+++ src/sql_type/object.rs
@@ -17,7 +17,6 @@ use std::cmp;
 use std::fmt;
 use std::mem::{self, MaybeUninit};
 use std::os::raw::c_char;
-use std::ptr;
 use std::sync::Arc;
 
 use crate::binding::*;
@@ -29,8 +28,10 @@ use crate::sql_type::OracleType;
 use crate::sql_type::ToSql;
 use crate::to_rust_str;
 use crate::util::write_literal;
+use crate::AssertSend;
 use crate::Connection;
 use crate::Context;
+use crate::DpiObject;
 use crate::DpiObjectAttr;
 use crate::DpiObjectType;
 use crate::Error;
@@ -77,12 +78,12 @@ unsafe fn release_dpi_data(data: &dpiData, native_type_num: u32) {
 /// Note: Methods in the type may be changed in future.
 pub struct Collection {
     conn: Conn,
-    pub(crate) handle: *mut dpiObject,
+    pub(crate) handle: DpiObject,
     objtype: ObjectType,
 }
 
 impl Collection {
-    pub(crate) fn new(conn: Conn, handle: *mut dpiObject, objtype: ObjectType) -> Collection {
+    pub(crate) fn new(conn: Conn, handle: DpiObject, objtype: ObjectType) -> Collection {
         Collection {
             conn,
             handle,
@@ -92,6 +93,10 @@ impl Collection {
 
     pub(crate) fn ctxt(&self) -> &Context {
         self.conn.ctxt()
+    }
+
+    fn handle(&self) -> *mut dpiObject {
+        self.handle.raw
     }
 
     /// Returns type information.
@@ -106,7 +111,7 @@ impl Collection {
     /// [OCICollSize()]: https://www.oracle.com/pls/topic/lookup?ctx=dblatest&id=GUID-B8F6665F-12F1-43DB-A27E-82A2A655D701
     pub fn size(&self) -> Result<i32> {
         let mut size = 0;
-        chkerr!(self.ctxt(), dpiObject_getSize(self.handle, &mut size));
+        chkerr!(self.ctxt(), dpiObject_getSize(self.handle(), &mut size));
         Ok(size)
     }
 
@@ -257,7 +262,7 @@ impl Collection {
         let mut exists = 0;
         chkerr!(
             self.ctxt(),
-            dpiObject_getFirstIndex(self.handle, &mut index, &mut exists)
+            dpiObject_getFirstIndex(self.handle(), &mut index, &mut exists)
         );
         if exists != 0 {
             Ok(index)
@@ -274,7 +279,7 @@ impl Collection {
         let mut exists = 0;
         chkerr!(
             self.ctxt(),
-            dpiObject_getLastIndex(self.handle, &mut index, &mut exists)
+            dpiObject_getLastIndex(self.handle(), &mut index, &mut exists)
         );
         if exists != 0 {
             Ok(index)
@@ -291,7 +296,7 @@ impl Collection {
         let mut exists = 0;
         chkerr!(
             self.ctxt(),
-            dpiObject_getNextIndex(self.handle, index, &mut next, &mut exists)
+            dpiObject_getNextIndex(self.handle(), index, &mut next, &mut exists)
         );
         if exists != 0 {
             Ok(next)
@@ -308,7 +313,7 @@ impl Collection {
         let mut exists = 0;
         chkerr!(
             self.ctxt(),
-            dpiObject_getPrevIndex(self.handle, index, &mut prev, &mut exists)
+            dpiObject_getPrevIndex(self.handle(), index, &mut prev, &mut exists)
         );
         if exists != 0 {
             Ok(prev)
@@ -322,7 +327,7 @@ impl Collection {
         let mut exists = 0;
         chkerr!(
             self.ctxt(),
-            dpiObject_getElementExistsByIndex(self.handle, index, &mut exists)
+            dpiObject_getElementExistsByIndex(self.handle(), index, &mut exists)
         );
         Ok(exists != 0)
     }
@@ -349,7 +354,7 @@ impl Collection {
             chkerr!(
                 self.ctxt(),
                 dpiObject_getElementValueByIndex(
-                    self.handle,
+                    self.handle(),
                     index,
                     native_type_num,
                     sql_value.data
@@ -370,7 +375,7 @@ impl Collection {
         chkerr!(
             self.ctxt(),
             dpiObject_setElementValueByIndex(
-                self.handle,
+                self.handle(),
                 index,
                 sql_value.native_type_num(),
                 sql_value.data
@@ -387,7 +392,7 @@ impl Collection {
         sql_value.set(value)?;
         chkerr!(
             self.ctxt(),
-            dpiObject_appendElement(self.handle, sql_value.native_type_num(), sql_value.data)
+            dpiObject_appendElement(self.handle(), sql_value.native_type_num(), sql_value.data)
         );
         Ok(())
     }
@@ -398,7 +403,7 @@ impl Collection {
     pub fn remove(&mut self, index: i32) -> Result<()> {
         chkerr!(
             self.ctxt(),
-            dpiObject_deleteElementByIndex(self.handle, index)
+            dpiObject_deleteElementByIndex(self.handle(), index)
         );
         Ok(())
     }
@@ -408,21 +413,14 @@ impl Collection {
     /// If the number of of elements to trim exceeds the current size
     /// of the collection an error is returned.
     pub fn trim(&mut self, len: usize) -> Result<()> {
-        chkerr!(self.ctxt(), dpiObject_trim(self.handle, len as u32));
+        chkerr!(self.ctxt(), dpiObject_trim(self.handle(), len as u32));
         Ok(())
     }
 }
 
 impl Clone for Collection {
     fn clone(&self) -> Collection {
-        unsafe { dpiObject_addRef(self.handle) };
-        Collection::new(self.conn.clone(), self.handle, self.objtype.clone())
-    }
-}
-
-impl Drop for Collection {
-    fn drop(&mut self) {
-        let _ = unsafe { dpiObject_release(self.handle) };
+        Collection::new(self.conn.clone(), self.handle.clone(), self.objtype.clone())
     }
 }
 
@@ -481,6 +479,8 @@ impl fmt::Debug for Collection {
     }
 }
 
+impl AssertSend for Collection {}
+
 /// Oracle-specific object data type
 ///
 /// ```no_run
@@ -510,12 +510,12 @@ impl fmt::Debug for Collection {
 /// Note: Methods in the type may be changed in future.
 pub struct Object {
     conn: Conn,
-    pub(crate) handle: *mut dpiObject,
+    pub(crate) handle: DpiObject,
     objtype: ObjectType,
 }
 
 impl Object {
-    pub(crate) fn new(conn: Conn, handle: *mut dpiObject, objtype: ObjectType) -> Object {
+    pub(crate) fn new(conn: Conn, handle: DpiObject, objtype: ObjectType) -> Object {
         Object {
             conn,
             handle,
@@ -525,6 +525,10 @@ impl Object {
 
     pub(crate) fn ctxt(&self) -> &Context {
         self.conn.ctxt()
+    }
+
+    pub(crate) fn handle(&self) -> *mut dpiObject {
+        self.handle.raw
     }
 
     /// Returns type information.
@@ -561,7 +565,7 @@ impl Object {
             chkerr!(
                 self.ctxt(),
                 dpiObject_getAttributeValue(
-                    self.handle,
+                    self.handle(),
                     attr.handle.raw(),
                     native_type_num,
                     sql_value.data
@@ -591,7 +595,7 @@ impl Object {
         chkerr!(
             self.ctxt(),
             dpiObject_setAttributeValue(
-                self.handle,
+                self.handle(),
                 attrtype.handle.raw(),
                 sql_value.native_type_num(),
                 sql_value.data
@@ -603,14 +607,7 @@ impl Object {
 
 impl Clone for Object {
     fn clone(&self) -> Object {
-        unsafe { dpiObject_addRef(self.handle) };
-        Object::new(self.conn.clone(), self.handle, self.objtype.clone())
-    }
-}
-
-impl Drop for Object {
-    fn drop(&mut self) {
-        let _ = unsafe { dpiObject_release(self.handle) };
+        Object::new(self.conn.clone(), self.handle.clone(), self.objtype.clone())
     }
 }
 
@@ -661,6 +658,8 @@ impl fmt::Debug for Object {
         write!(f, "))")
     }
 }
+
+impl AssertSend for Object {}
 
 /// Type information about Object or Collection data type
 ///
@@ -782,10 +781,10 @@ impl ObjectType {
             )));
         }
         let conn = &self.internal.conn;
-        let mut handle = ptr::null_mut();
+        let mut handle = DpiObject::null();
         chkerr!(
             conn.ctxt(),
-            dpiObjectType_createObject(self.internal.handle.raw(), &mut handle)
+            dpiObjectType_createObject(self.internal.handle.raw(), &mut handle.raw)
         );
         Ok(Object::new(conn.clone(), handle, self.clone()))
     }
@@ -800,10 +799,10 @@ impl ObjectType {
             )));
         }
         let conn = &self.internal.conn;
-        let mut handle = ptr::null_mut();
+        let mut handle = DpiObject::null();
         chkerr!(
             conn.ctxt(),
-            dpiObjectType_createObject(self.internal.handle.raw(), &mut handle)
+            dpiObjectType_createObject(self.internal.handle.raw(), &mut handle.raw)
         );
         Ok(Collection::new(conn.clone(), handle, self.clone()))
     }